refactor: Cleanup, moved to ui directory
