+++ packages/apps/spaces/components/shared/location/Location.tsx
@@ -7,7 +7,7 @@ import {
   SelectMenu,
   SelectInput,
   SelectWrapper,
-} from '@spaces/ui-kit/select';
+} from '@spaces/ui-kit/../../../ui/form/select';
 
 export const Location: React.FC<{
   locationId: string;
